Make AtmosphereExtinctionThreshold float constexpr

+++ src/celengine/shadermanager.cpp
@@ -3692,9 +3692,9 @@ CelestiaGLProgram::setAtmosphereParameters(const Atmosphere& atmosphere,
     // fallse off exponentially with height above the planet's surface, so the actual
     // radius is infinite. That's a bit impractical, so well just render the portion
     // out to the point where the density is some fraction of the surface density.
-    float skySphereRadius = atmPlanetRadius + -atmosphere.mieScaleHeight * (float) log(AtmosphereExtinctionThreshold);
+    float skySphereRadius = atmPlanetRadius + -atmosphere.mieScaleHeight * log(AtmosphereExtinctionThreshold);
 
-    float tMieCoeff        = atmosphere.mieCoeff * objRadius;
+    float tMieCoeff           = atmosphere.mieCoeff * objRadius;
     Vector3f tRayleighCoeff   = atmosphere.rayleighCoeff * objRadius;
     Vector3f tAbsorptionCoeff = atmosphere.absorptionCoeff * objRadius;
